Fixup: Reserve more PEBs to avoid -ENOSPC when filling up device
Currently, the accounting does not consider the situation when almost all LEBs
are completely filled up, fully cosolidated and the remaining few LEBs are
written to. More specifically, if lebs_per_clebs = 2 (MLC) and all LEBs except
1 are fully filled up and consolidate, a write to the last free LEB will result
in -ENOSPC since we are missing a PEB to write to.

To fix that, we reserve an additional lebs_per_clebs-1 PEBs. (Yes, this does waste
some PEBs, but these will on the other hand be available for wear leveling etc.)

Signed-off-by: David Gstir <david@sigma-star.at>
Signed-off-by: Richard Weinberger <richard@nod.at>