ubi: Finally resolve VTBL accounting hack
We used to cheat and marked the initial VTBL LEBs
always as full. Upon volume creation new VTBLs are created
and these are not always full.
No consolidiation happened and we thought that the accounting
is broken.
Therefore we always reserved two PEBs for the VTBLs which made
things work again. But it covered only the real problem.

Remove that hack and mark the VTBL only full when they are.
That way they can be consolidated like any other LEB.

Signed-off-by: Richard Weinberger <richard@nod.at>